feat: add ERC-6909 Token and Supply extension (#777)
<!--
Thank you for your interest in contributing to OpenZeppelin!

Consider opening an issue for discussion prior to submitting a PR. New
features will be merged faster if they were first discussed and designed
with the team.

Describe the changes introduced in this pull request. Include any
context necessary for understanding the PR's purpose.
-->

<!-- Fill in with issue number -->

#### PR Checklist

<!--
Before merging the pull request all of the following must be completed.
Feel free to submit a PR or Draft PR even if some items are pending.
Some of the items may not apply.
-->

- [x] Tests
- [x] Documentation
- [x] Changelog

# ERC6909 Implementation in rust-contracts-stylus

## Summary

This PR implements the
[ERC-6909](https://github.com/OpenZeppelin/openzeppelin-contracts/blob/v5.3.0/contracts/token/ERC6909/draft-ERC6909.sol)
token and [Supply
extension](https://github.com/OpenZeppelin/openzeppelin-contracts/blob/v5.3.0/contracts/token/ERC6909/extensions/draft-ERC6909TokenSupply.sol)
in Rust for Arbitrum Stylus.

The following implementation is provided in this PR:

- ERC-6909 + unit tests with motsu
- Supply Extension + unit tests with motsu
- Example usage of ERC-6909 and Supply Extension
- Integration Tests for ERC-6909 and Supply Extension
- Benchmark for both ERC6909 and Supply Extension

### Comments
Integration tests and the benchmarks inside
[rust-contracts-stylus](https://github.com/OpenZeppelin/rust-contracts-stylus),
which use
[nitro-testnode](https://github.com/OffchainLabs/nitro-testnode), didn't
work with the M3 Pro Chip (would be an issue related to Docker). To
solve the problem, firstly, [nitro-devnode] is tried to use. However, I
observed that it is not suitable for the `e2e::test` workflow. So, an
LTS Ubuntu machine is rented on AWS to run the integration tests and the
benchmark.

For the Supply Extension implementation, `Deref` traits are implemented.
Using `Deref` allows direct access to Erc6909 methods, improving
readability and maintainability. However, deferencing operations might
increase the gas. Designing Supply Token Extension in this way increases
the abstraction and makes it easy to use for users who would like to use
the Supply Extension with ERC-6909 Token Contract.

### Future Improvements

Currently, this PR does not include other ERC-6909 Extensions,
specifically ContentURI and Metadata. Their example usage and the
benchmarks should be added to the repository.

The unit tests that differentiate the `id` should be added.

### Benchmark

Benchmarks for ERC-6909 can be seen as follows:

```
| Contract::function                                                        | WASM Opt & Cached | Cached | Not Cached |
| ------------------------------------------------------------------------- | ----------------- | ------ | ---------- |
| Erc6909::mint(address,uint256,uint256)                                    |             27925 |  29307 |      45089 |
| Erc6909::balanceOf(address,uint256)                                       |              5101 |   6448 |      22230 |
| Erc6909::allowance(address,address,uint256)                               |              5875 |   7228 |      23010 |
| Erc6909::isOperator(address,address)                                      |              5322 |   6666 |      22448 |
| Erc6909::setOperator(address,bool)                                        |             27045 |  28433 |      44215 |
| Erc6909::transfer(address,uint256,uint256)                                |             34364 |  35786 |      51568 |
| Erc6909::approve(address,uint256,uint256)                                 |             27939 |  29332 |      45114 |
| Erc6909::transferFrom(address,address,uint256,uint256)                    |             20503 |  21943 |      37725 |
| Erc6909::burn(address,uint256,uint256)                                    |             11495 |  12893 |      28675 |

```

---------

Co-authored-by: Daniel Bigos <daniel.bigos@openzeppelin.com>
Co-authored-by: Alisander Qoshqosh <qalisander@gmail.com>
Co-authored-by: Daniel Bigos <daniel.bigos96@gmail.com>
Co-authored-by: Nenad <xinef.it@gmail.com>